preparend alert-replacement moving more modules out of the main file replacing all calls to window.alert with idrinth.alert so further replacements are easier
