+++ stable.js
@@ -20,160 +20,6 @@ var idrinth = {
     escapeRegExp: function ( str ) {
         return str.replace ( /[\-\[\]\/\{\}\(\)\*\+\?\.\\\^\$\|]/g, "\\$&" );
     },
-    war: {
-        from: null,
-        to: null,
-        element: null,
-        oc: function () {
-            var c = idrinth.war.element.getAttribute ( 'class' );
-            if ( c.indexOf ( 'small-window' ) > -1 ) {
-                c = c.replace ( 'small-window' );
-            } else {
-                c = c + ' small-window';
-            }
-            idrinth.war.element.setAttribute ( 'class', c );
-        },
-        getData: function () {
-            idrinth.runAjax (
-                    "https://dotd.idrinth.de/" + ( idrinth.settings.isWorldServer ? "world-" : "" ) + idrinth.platform + "/war-service/" + idrinth.war.warRaids2Join () + "/"+Date.now ()+"/",
-                    idrinth.war.updateData,
-                    function () {
-                        window.setTimeout ( idrinth.war.getData, 1000 );
-                    },
-                    function () {
-                        window.setTimeout ( idrinth.war.getData, 2000 );
-                    },
-                    idrinth.raids.knowRaids ()
-                    );
-        },
-        warRaids2Join: function () {
-            var list = [ ];
-            for (var input in idrinth.war.element.getElementsByTagName ( 'input' )) {
-                if ( idrinth.war.element.getElementsByTagName ( 'input' )[input].checked ) {
-                    list.push ( idrinth.war.element.getElementsByTagName ( 'input' )[input].getAttribute ( 'data-id' ) );
-                }
-            }
-            if ( list.length > 0 ) {
-                return list.join ( ',' );
-            }
-            return '_';
-        },
-        updateData: function ( data ) {
-            function process ( data ) {
-                function addRaids ( raids ) {
-                    for (var key in raids) {
-                        if ( idrinth.raids.joined[key] === undefined && idrinth.raids.list[key] === undefined ) {
-                            idrinth.raids.list[key] = raids[key];
-                        }
-                    }
-                }
-                function updateBoss ( data, element ) {
-                    while ( element.getElementsByTagName ( 'td' )[3].firstChild ) {
-                        element.getElementsByTagName ( 'td' )[3].removeChild ( element.getElementsByTagName ( 'td' )[3].firstChild );
-                    }
-                    if(data.magics!==null&&data.magics!=='') {
-                        var tmp = data.magics.split ( ',' );
-                        for (var key = 0; key < tmp.length; key++) {
-                            element.getElementsByTagName ( 'td' )[3].appendChild (
-                                    idrinth.ui.buildElement ( {
-                                        type: 'img',
-                                        attributes: [
-                                            { name: 'src', value: 'https://dotd.idrinth.de/static/magic-image-service/' + data.magics.split ( ',' )[key] + '/' },
-                                            { name: 'width', value: '20' }
-                                        ] } )
-                                    );
-                        }
-                    }
-                    element.getElementsByTagName ( 'td' )[0].setAttribute ( "class", 'traffic ' + ( data.amount < 90 ? 'yes' : ( data.amount > 110 ? 'no' : 'maybe' ) ) );
-                    element.getElementsByTagName ( 'td' )[0].setAttribute ( "title", data.amount + '/100' );
-                    element.getElementsByTagName ( 'td' )[0].firstChild.innerHTML = ( data.amount < 90 ? 'yes' : ( data.amount > 110 ? 'no' : 'maybe' ) );
-                }
-                function newBoss ( data, boss ) {
-                    var magics = [ ];
-                    if(data.magics!==null&&data.magics!=='') {
-                        var tmp = data.magics.split ( ',' );
-                        for (var key = 0; key < tmp.length; key++) {
-                            magics.push ( {
-                                type: 'img',
-                                attributes: [
-                                    { name: 'src', value: 'https://dotd.idrinth.de/static/magic-image-service/' + data.magics.split ( ',' )[key] + '/' },
-                                    { name: 'width', value: '20' }
-                                ] }
-                            );
-                        }
-                    }
-                    idrinth.war.element.childNodes[1].appendChild ( idrinth.ui.buildElement (
-                            {
-                                type: 'tr',
-                                id: 'idrinth-war-' + boss,
-                                children: [
-                                    {
-                                        type: 'td',
-                                        css: 'traffic ' + ( data.amount < 90 ? 'yes' : ( data.amount > 110 ? 'no' : 'maybe' ) ),
-                                        children: [ { type: 'span', content: ( data.amount < 90 ? 'yes' : ( data.amount > 110 ? 'no' : 'maybe' ) ) } ],
-                                        attributes: [
-                                            { name: 'title', value: data.amount + '/100' }
-                                        ]
-                                    },
-                                    {
-                                        type: 'td',
-                                        content: data.name
-                                    },
-                                    {
-                                        type: 'td',
-                                        children: [
-                                            {
-                                                type: 'input',
-                                                attributes: [
-                                                    { name: 'type', value: 'checkbox' },
-                                                    { name: 'data-id', value: boss },
-                                                    { name: 'title', value: 'join ' + data.name }
-                                                ]
-                                            }
-                                        ]
-                                    },
-                                    {
-                                        type: 'td',
-                                        children: magics
-                                    }
-                                ]
-                            }
-                    ) );
-                }
-                if ( data === "" || data === "null" ) {
-                    return;
-                }
-                if ( data === "{}" ) {
-                    idrinth.war.element.setAttribute ( 'style', 'display:none' );
-                    while ( idrinth.war.element.childNodes[1].childNodes[1].firstChild ) {
-                        idrinth.war.element.childNodes[1].childNodes[1].removeChild ( idrinth.war.element.childNodes[1].firstChild.firstChild );
-                    }
-                    return;
-                }
-                idrinth.war.element.setAttribute ( 'style', '' );
-                var classes = idrinth.war.element.getAttribute ( 'class');
-                idrinth.war.element.setAttribute ( 'class', idrinth.settings.warBottom?classes+' bottom':classes.replace(/ bottom/g,'') );
-                try {
-                    data = JSON.parse ( data );
-                    if ( data.raids !== undefined ) {
-                        addRaids ( data.raids );
-                    }
-                    for (var boss in data.stats) {
-                        if ( document.getElementById ( 'idrinth-war-' + boss ) ) {
-                            updateBoss ( data.stats[boss], document.getElementById ( 'idrinth-war-' + boss ) );
-                        } else {
-                            newBoss ( data.stats[boss], boss );
-                        }
-                    }
-                } catch ( e ) {}
-            }
-            process ( data );
-            window.setTimeout ( idrinth.war.getData, 50000 + Math.random () % 20000 );
-        },
-        start: function () {
-            window.setTimeout ( idrinth.war.getData, 5000 );
-        }
-    },
     facebook: {
         popup: null,
         timeout: null,
@@ -221,7 +67,7 @@ var idrinth = {
         },
         join: function () {
             if ( idrinth.newgrounds.raids.length === 0 ) {
-                window.alert ( 'We\'re done! Have fun playing.' );
+                idrinth.alert ( 'We\'re done! Have fun playing.' );
                 return;
             }
             var frame = document.getElementById ( 'iframe_embed' ).getElementsByTagName ( 'iframe' )[0];
@@ -264,7 +110,7 @@ var idrinth = {
         idrinth.ui.removeElement ( 'idrinth-chat' );
         idrinth.ui.removeElement ( 'idrinth-war' );
         var sc = document.createElement ( 'script' );
-        sc.setAttribute ( 'src', 'https://dotd.idrinth.de/static/userscript/###RELOAD-VERSION###/' );
+        sc.setAttribute ( 'src', 'https://dotd.idrinth.de/static/userscript/' + Math.random ()+'/' );
         document.getElementsByTagName ( 'body' )[0].appendChild ( sc );
         window.setTimeout ( function () {
             idrinth = { };
@@ -375,260 +221,8 @@ var idrinth = {
             idrinth.ui.makeTierList ( result );
         }
     },
-    land: {
-        calculate: function () {
-            for (var key in idrinth.settings.land) {
-                idrinth.settings.land[key] = parseInt ( document.getElementById ( 'idrinth-land-' + key ).value );
-            }
-            var results = idrinth.settings.landMax ? idrinth.land.useUp () : idrinth.land.bestPrice ();
-            if(Object.keys (results).length===0) {
-                window.alert('You lack gold to buy any more buildings at the moment.');
-            }
-            idrinth.land.putResults ( results );
-        },
-        bestPrice: function () {
-            var factor = idrinth.settings.factor ? 10 : 1;
-            var results = { };
-            while ( idrinth.settings.land.gold >= 0 ) {
-                var min = null;
-                var minKey = null;
-                for (var key in idrinth.land.data) {
-                    if (
-                            min === null ||
-                            ( 10 + idrinth.settings.land[key] ) * idrinth.land.data[key].base / idrinth.land.data[key].perHour < min )
-                    {
-                        min = ( 10 + idrinth.settings.land[key] ) * idrinth.land.data[key].base / idrinth.land.data[key].perHour;
-                        minKey = key;
-                    }
-                }
-                if ( minKey === null ) {
-                    return results;
-                }
-                var price = ( 10 + idrinth.settings.land[minKey] ) * factor * idrinth.land.data[minKey].base / 10;
-                if ( price > idrinth.settings.land.gold ) {
-                    idrinth.land.putResults ( results );
-                    return;
-                }
-                idrinth.settings.land.gold = idrinth.settings.land.gold - price;
-                results[minKey] = ( results[minKey] === undefined ? 0 : results[minKey] ) + factor;
-                idrinth.settings.land[minKey] = idrinth.settings.land[minKey] + factor;
-            }
-            return results;
-        },
-        useUp: function () {
-            var factor = idrinth.settings.factor ? 10 : 1;
-            var results = { };
-            while ( idrinth.settings.land.gold >= 0 ) {
-                var min = null;
-                var minKey = null;
-                for (var key in idrinth.land.data) {
-                    if (
-                            ( min === null ||
-                                    ( 10 + idrinth.settings.land[key] ) * idrinth.land.data[key].base / idrinth.land.data[key].perHour < min )
-                            && ( 10 + idrinth.settings.land[key] ) * factor * idrinth.land.data[key].base / 10 <= idrinth.settings.land.gold )
-                    {
-                        min = ( 10 + idrinth.settings.land[key] ) * idrinth.land.data[key].base / idrinth.land.data[key].perHour;
-                        minKey = key;
-                    }
-                }
-                if ( minKey === null ) {
-                    return results;
-                }
-                idrinth.settings.land.gold = idrinth.settings.land.gold - ( 10 + idrinth.settings.land[minKey] ) * factor * idrinth.land.data[minKey].base / 10;
-                results[minKey] = ( results[minKey] === undefined ? 0 : results[minKey] ) + factor;
-                idrinth.settings.land[minKey] = idrinth.settings.land[minKey] + factor;
-            }
-            return results;
-        },
-        putResults: function ( results ) {
-            for (var key in results) {
-                document.getElementById ( 'idrinth-land-' + key ).value = idrinth.settings.land[key];
-                document.getElementById ( 'idrinth-land-' + key ).parentNode.nextSibling.innerHTML = '+' + results[key];
-            }
-            document.getElementById ( 'idrinth-land-gold' ).value = idrinth.settings.land.gold;
-            idrinth.settings.save ();
-        },
-        data: {
-            cornfield: { perHour: 100, base: 4000 },
-            stable: { perHour: 300, base: 15000 },
-            barn: { perHour: 400, base: 25000 },
-            store: { perHour: 700, base: 50000 },
-            pub: { perHour: 900, base: 75000 },
-            inn: { perHour: 1200, base: 110000 },
-            tower: { perHour: 2700, base: 300000 },
-            fort: { perHour: 4500, base: 600000 },
-            castle: { perHour: 8000, base: 1200000 }
-        }
-    },
-    names: {
-        load: function ( added) {
-            'use strict';
-            idrinth.runAjax (
-                    'https://dotd.idrinth.de/' + idrinth.platform + '/users-service/'+added,
-                    function ( text ) {
-                        idrinth.names.import ( text );
-                    }
-                );
-        },
-        import: function ( data ) {
-            'use strict';
-            data = JSON.parse ( data );
-            if ( !data ) {
-                return;
-            }
-            if ( data.users ) {
-                idrinth.names.users = data.users;
-            }
-            if ( data.guilds ) {
-                idrinth.names.guilds = data.guilds;
-            }
-            if ( data.classes ) {
-                idrinth.names.classes = data.classes;
-            }
-        },
-        run: function ( ) {
-            'use strict';
-            try{
-                if ( idrinth.names.counter % 300 === 0 || Object.keys ( idrinth.names.users ).length === 0) {
-                    idrinth.names.load ( Object.keys ( idrinth.names.classes ).length === 0?'init/':'get/' );
-                } else if ( Object.keys ( idrinth.names.users ).length > 0 ) {
-                    idrinth.names.add ( );
-                }
-            } catch(e) {
-                console.log (e);
-            }
-            idrinth.names.counter = idrinth.names.counter + 1;
-            idrinth.names.ownTimeout = window.setTimeout ( function ( ) {
-                idrinth.names.run ( );
-            }, 6666 );
-        },
-        users: { },
-        classes: { },
-        guilds: { },
-        ownTimeout: null,
-        add: function ( ) {
-            'use strict';
-            var el = document.getElementsByClassName ( 'username' );
-            var name = '';
-            var count = 0;
-            for (count = el.length - 1; count >= 0; count--) {
-                try {
-                    name = idrinth.names.parse ( el[count] );
-                    if ( el[count].getAttribute ( 'onmouseover' ) !== 'idrinth.ui.showTooltip()(this);' && idrinth.ui.childOf ( el[count], 'chat_message_window' ) ) {
-                        el[count].setAttribute ( 'onmouseover', 'idrinth.ui.showTooltip(this);' );
-                    }
-                    if ( !idrinth.names.users[name.toLowerCase ( )] && name.length > 0 ) {
-                        idrinth.names.users[name.toLowerCase()] = { };
-                        idrinth.runAjax (
-                           'https://dotd.idrinth.de/' + idrinth.platform + '/users-service/add/'+encodeURIComponent(name)+'/'
-                        );
-                    }
-                } catch ( e ) {
-                }
-            }
-        },
-        parse: function ( element ) {
-            'use strict';
-            if ( element.getAttribute ( 'dotdxname' ) ) {
-                return ( element.getAttribute ( 'dotdxname' ) );
-            }
-            if ( element.getAttribute ( 'username' ) ) {
-                return ( element.getAttribute ( 'username' ) );
-            }
-            return ( element.innerHTML ).replace ( /(<([^>]+)>)/ig, "" );
-        },
-        start: function ( ) {
-            'use strict';
-            if ( idrinth.platform === 'kongregate' ) {
-                idrinth.names.ownTimeout = window.setTimeout ( function ( ) {
-                    idrinth.names.run ( );
-                }, 10000 );
-            }
-        },
-        counter: 0
-    },
     platform: '',
     realSite: '',
-    settings: {
-        raids: false,
-        favs: '',
-        factor: true,
-        moveLeft: false,
-        minimalist: false,
-        names: true,
-        timeout: 5000,
-        loadtime: 5000,
-        windows: 3,
-        warBottom:false,
-        landMax: true,
-        chatting: true,
-        chatuser: '',
-        newgroundLoad: 30,
-        chatpass: '',
-        isWorldServer: false,
-        alarmTime: '8:0',
-        alarmActive: false,
-        land: {
-            cornfield: 0,
-            stable: 0,
-            barn: 0,
-            store: 0,
-            pub: 0,
-            inn: 0,
-            tower: 0,
-            fort: 0,
-            castle: 0,
-            gold: 0
-        },
-        save: function ( ) {
-            'use strict';
-            if ( window.localStorage ) {
-                for (var key in idrinth.settings) {
-                    if ( key === 'land' ) {
-                        for (var building in idrinth.settings.land) {
-                            if ( typeof idrinth.settings[key] !== 'function' ) {
-                                window.localStorage.setItem ( 'idrinth-dotd-land-' + building, idrinth.settings.land[building] );
-                            }
-                        }
-                    } else if ( typeof idrinth.settings[key] !== 'function' ) {
-                        window.localStorage.setItem ( 'idrinth-dotd-' + key, idrinth.settings[key] );
-                    }
-                }
-            }
-        },
-        change: function ( field, value ) {
-            'use strict';
-            idrinth.settings[field] = value;
-            idrinth.settings.save ( );
-        },
-        start: function ( ) {
-            if ( window.localStorage ) {
-                'use strict';
-                for (var key in idrinth.settings) {
-                    if ( key === 'land' ) {
-                        for (var building in idrinth.settings.land) {
-                            if ( typeof idrinth.settings[key] !== 'function' ) {
-                                var tmp = window.localStorage.getItem ( 'idrinth-dotd-land-' + building );
-                                if ( tmp ) {
-                                    idrinth.settings.land[building] = tmp;
-                                }
-                            }
-                        }
-                    } else if ( typeof idrinth.settings[key] !== 'function' ) {
-                        var tmp = window.localStorage.getItem ( 'idrinth-dotd-' + key );
-                        if ( tmp ) {
-                            if ( tmp === 'false' ) {
-                                tmp = false;
-                            } else if ( tmp === 'true' ) {
-                                tmp = true;
-                            }
-                            idrinth.settings[key] = tmp;
-                        }
-                    }
-                }
-            }
-        }
-    },
     log: function ( string ) {
         'use strict';
         console.log ( '[IDotDS] ' + string );
@@ -1715,7 +1309,7 @@ var idrinth = {
                         value: 'stylesheet'
                     }, {
                         name: 'href',
-                        value: 'https://dotd.idrinth.de###PATH###/script-styles.css?' + idrinth.version
+                        value: 'https://dotd.idrinth.de###PATH###/script-styles.css'
                     } ]
             } ) );
             if ( idrinth.realSite === 'kongregate' ) {
@@ -1724,370 +1318,6 @@ var idrinth = {
             idrinth.ui.buildBasis.do ( );
             idrinth.war.element = idrinth.ui.buildBasis.war ();
             idrinth.ui.body.appendChild ( idrinth.war.element );
-        }
-    },
-    user: {
-        token: '',
-        id: 0,
-        name: '',
-        identifier: '',
-        start: function ( ) {
-            'use strict';
-            if ( idrinth.realSite === 'kongregate' ) {
-                idrinth.user.name = active_user._attributes.get ( 'username' );
-                idrinth.user.token = active_user.gameAuthToken ( );
-                idrinth.user.id = active_user.id ( );
-            } else if ( idrinth.realSite === 'newgrounds' ) {
-                idrinth.user.name = Cookies.get ( 'NG_GG_username' );
-            } else if ( idrinth.realSite === 'armorgames' ) {
-                var ag = document.getElementById ( 'gamefilearea' ).children[0].src.match ( /^.+user_id=([a-f\d]{32})&auth_token=([a-f\d]{32}).+$/ );
-                idrinth.user.name = window.u_name;
-                idrinth.user.id = ag[1];
-                idrinth.user.token = ag[2];
-            }
-            window.setTimeout ( idrinth.user.sendAlive, 20000 );
-        },
-        sendAlive: function () {
-            function guid () {
-                //from http://stackoverflow.com/a/105074
-                function s4 () {
-                    return Math.floor ( ( 1 + Math.random () ) * 0x10000 ).toString ( 36 );
-                }
-                return s4 () + '-' +
-                        s4 () + s4 () + '-' +
-                        s4 () + s4 () + s4 () + '-' +
-                        s4 () + s4 () + s4 () + s4 () + '-' +
-                        s4 () + s4 () + s4 () + s4 () + s4 () + '-' +
-                        s4 () + s4 () + s4 () + s4 () + s4 () + s4 ();
-            }
-            if ( window.localStorage ) {
-                idrinth.user.identifier = window.localStorage.getItem ( 'idrinth-dotd-uuid' );
-                if ( !idrinth.user.identifier || idrinth.user.identifier === '' || idrinth.user.identifier === null || !idrinth.user.identifier.match (/^[a-z0-9]{4}-[a-z0-9]{8}-[a-z0-9]{12}-[a-z0-9]{16}-[a-z0-9]{20}-[a-z0-9]{24}$/) ) {
-                    idrinth.user.identifier = guid ();
-                }
-                window.localStorage.setItem ( 'idrinth-dotd-uuid', idrinth.user.identifier );
-                idrinth.runAjax ( 'https://dotd.idrinth.de/' +
-                        ( idrinth.settings.isWorldServer ? 'world-' : '' ) + idrinth.platform +
-                        '/i-am-alive/' + idrinth.user.identifier + '/' );
-            }
-        }
-    },
-    raids: {
-        script: null,
-        list: { },
-        joined: { },
-        interval: null,
-        requested: 0,
-        restartInterval: function () {
-            try {
-                window.clearInterval ( idrinth.raids.interval );
-            } catch ( e ) {
-            }
-            ;
-            idrinth.raids.interval = window.setInterval ( function ( ) {
-                idrinth.raids.join.process ( );
-            }, 1500 );
-        },
-        getImportLink: function ( toImport ) {
-            return 'https://dotd.idrinth.de/' + ( idrinth.settings.isWorldServer ? 'world-' : '' ) + idrinth.platform +
-                    '/raid-service/' + ( toImport === '' ? '_' : toImport ) + '/';
-        },
-        import: function ( ) {
-            'use strict';
-            if ( !idrinth.platform ) {
-                return;
-            }
-            idrinth.runAjax (
-                    idrinth.raids.getImportLink ( idrinth.settings.raids ? idrinth.settings.favs : '-1' ),
-                    function ( text ) {
-                        idrinth.raids.importProcess ( text );
-                    },
-                    function ( ) {
-                    },
-                    function ( ) {
-                    },
-                    idrinth.raids.knowRaids ()
-                    );
-        },
-        knowRaids: function () {
-            var data = '';
-            if ( Object.keys ( idrinth.raids.joined ).length > 0 ) {
-                data = data + ',' + ( Object.keys ( idrinth.raids.joined ) ).join ();
-            }
-            if ( Object.keys ( idrinth.raids.list ).length > 0 ) {
-                data = data + ',' + ( Object.keys ( idrinth.raids.list ) ).join ();
-            }
-            return data.replace ( /^,|,$/, '' );
-        },
-        importManually: function ( all ) {
-            'use strict';
-            if ( !idrinth.platform ) {
-                return;
-            }
-            idrinth.runAjax (
-                    idrinth.raids.getImportLink ( all ? '_' : idrinth.settings.favs ),
-                    function ( text ) {
-                        idrinth.raids.importProcess ( text );
-                    },
-                    function ( ) {
-                    },
-                    function ( ) {
-                    },
-                    idrinth.raids.knowRaids ()
-                    );
-        },
-        importProcess: function ( responseText ) {
-            'use strict';
-            var list = JSON.parse ( responseText );
-            for (var key in list) {
-                if ( list[key].delete ) {
-                    if ( idrinth.raids.list[key] ) {
-                        try {
-                            delete idrinth.raids.list[key];
-                        } catch ( e1 ) {
-                        }
-                        try {
-                            delete idrinth.raids.joined[key];
-                        } catch ( e2 ) {
-                        }
-                        try {
-                            idrinth.ui.removeElement ( 'idrinth-raid-link-' + key );
-                        } catch ( e ) {
-                        }
-                    }
-                } else if ( idrinth.raids.list[key] === undefined ) {
-                    idrinth.raids.list[key] = list[key];
-                }
-            }
-        },
-        clearAll: function () {
-            try {
-                window.clearInterval ( idrinth.raids.interval );
-            } catch ( e ) {
-            }
-            ;
-            while ( document.getElementById ( "idrinth-raid-link-list" ).firstChild ) {
-                idrinth.ui.removeElement ( document.getElementById ( "idrinth-raid-link-list" ).firstChild.id );
-            }
-            idrinth.raids.list = { };
-            idrinth.raids.joined = { };
-            idrinth.raids.restartInterval ();
-        },
-        join: {
-            data: {
-                prefix: null,
-                makePrefix: function () {
-                    if ( idrinth.raids.join.data.prefix === null ) {
-                        if ( idrinth.realSite === 'armorgames' ) {
-                            idrinth.raids.join.data.prefix = 'http://50.18.191.15/armor/raidjoin.php?user_id=' + idrinth.user.id +
-                                    '&auth_token=' + idrinth.user.token + '&';
-                        }
-                        if ( idrinth.realSite === 'kongregate' ) {
-                            idrinth.raids.join.data.prefix = 'http://50.18.191.15/kong/raidjoin.php?kongregate_username=' + idrinth.user.name +
-                                    '&kongregate_user_id=' + idrinth.user.id + '&kongregate_game_auth_token=' +
-                                    idrinth.user.token + '&';
-                        }
-                        if ( idrinth.realSite === 'newgrounds' ) {
-                            idrinth.raids.join.data.prefix = 'https://newgrounds.com/portal/view/609826?';
-                        }
-                        if ( idrinth.realSite === 'dawnofthedragons' || idrinth.realSite === 'facebook' ) {
-                            idrinth.raids.join.data.prefix = 'https://web1.dawnofthedragons.com/live_iframe/raidjoin.php?';
-                        }
-                    }
-                    return idrinth.raids.join.data.prefix;
-                },
-                tag: { armorgames: 'ar_', kongregate: 'kv_', newgrounds: 'ng_', facebook: '', dawnofthedragons: '' }
-            },
-            servers: {
-                getServerLink: function ( key ) {
-                    function build ( data, tag, prefix ) {
-                        try {
-                            return prefix +
-                                    tag + 'action_type=raidhelp&' +
-                                    tag + 'raid_id=' + data.raidId + '&' +
-                                    tag + 'difficulty=' + data.difficulty + '&' +
-                                    tag + 'hash=' + data.hash +
-                                    ( idrinth.settings.isWorldServer ? '&'+tag+'serverid=' + 2 : '' );
-                        } catch ( e1 ) {
-                            return false;
-                        }
-                    }
-                    var prefix = idrinth.raids.join.data.makePrefix ();
-                    var tag = idrinth.raids.join.data.tag[idrinth.realSite];
-                    if ( idrinth.raids.list[key] ) {
-                        var link = build ( idrinth.raids.list[key], tag, prefix );
-                        if ( link ) {
-                            return link;
-                        }
-                    }
-                    if ( idrinth.raids.joined[key] ) {
-                        var link = build ( idrinth.raids.joined[key], tag, prefix );
-                        if ( link ) {
-                            return link;
-                        }
-                    }
-                    return prefix;
-                },
-                kongregate: function ( key ) {
-                    'use strict';
-                    idrinth.raids.join.byAjax.do ( key, idrinth.raids.join.servers.getServerLink ( key ) );
-                },
-                armorgames: function ( key ) {
-                    'use strict';
-                    idrinth.raids.join.byAjax.do ( key, idrinth.raids.join.servers.getServerLink ( key ) );
-                },
-                postLink: function ( key ) {
-                    'use strict';
-                    if ( !document.getElementById ( 'idrinth-raid-link-' + key ) ) {
-                        var span = document.createElement ( 'span' );
-                        span.id = 'idrinth-raid-link-' + key;
-                        span.setAttribute ( 'data-clipboard-text', idrinth.raids.join.servers.getServerLink ( key ) );
-                        span.appendChild ( document.createTextNode ( idrinth.raids.list[key].name + '\'s ' + idrinth.raids.list[key].raid ) );
-                        document.getElementById ( 'idrinth-raid-link-list' ).appendChild ( span );
-                    }
-                },
-                facebook: function ( key ) {
-                    'use strict';
-                    idrinth.raids.join.byFrame.do ( key, idrinth.raids.join.servers.getServerLink ( key ) );
-                },
-                dawnofthedragons: function ( key ) {
-                    'use strict';
-                    idrinth.raids.join.byFrame.do ( key, idrinth.raids.join.servers.getServerLink ( key ) );
-                }
-            },
-            byAjax: {
-                'do': function ( key, link ) {
-                    'use strict';
-                    idrinth.runAjax (
-                            link,
-                            function ( ) {
-                                idrinth.raids.join.messages.success ( key );
-                            },
-                            function ( ) {
-                                idrinth.raids.join.messages.failed ( key );
-                            },
-                            function ( ) {
-                                idrinth.raids[key].joined = false;
-                                idrinth.raids.join.messages.failed ( key );
-                            }
-                    );
-                }
-            },
-            messages: {
-                writeToGui: function ( string ) {
-                    var li = document.createElement ( 'li' );
-                    var ul = document.getElementById ( 'idrinth-joined-raids' ).getElementsByTagName ( 'ul' )[0];
-                    li.appendChild ( document.createTextNode ( ( new Date () ).toLocaleTimeString () + ' ' + string ) );
-                    if ( !ul.firstChild ) {
-                        ul.appendChild ( li );
-                    } else {
-                        if ( ul.childNodes.length > 9 ) {
-                            ul.removeChild ( ul.lastChild );
-                        }
-                        ul.insertBefore ( li, ul.firstChild );
-                    }
-                },
-                log: function ( string ) {
-                    idrinth.log ( string );
-                    idrinth.raids.join.messages.writeToGui ( string );
-                },
-                success: function ( key ) {
-                    'use strict';
-                    idrinth.raids.join.messages.log ( 'Joined ' + idrinth.raids.list[key].name + '\'s ' + idrinth.raids.list[key].raid );
-                    try {
-                        idrinth.ui.removeElement ( 'idrinth-raid-link-' + key );
-                    } catch ( e ) {
-                    }
-                    try {
-                        idrinth.raids.joined[key] = idrinth.raids.list[key];
-                        delete idrinth.raids.list[key];
-                    } catch ( e1 ) {
-                    }
-                },
-                failed: function ( key ) {
-                    'use strict';
-                    idrinth.raids.join.messages.log ( 'Could not join ' + idrinth.raids.list[key].name + '\'s ' + idrinth.raids.list[key].raid );
-                },
-                trying: function ( key ) {
-                    'use strict';
-                    eval ( "window.setTimeout(function(){try{idrinth.ui.removeElement ('idrinth-raid-link-'+key);}catch(e){}},300)" );
-                    if ( idrinth.raids.list[key] ) {
-                        idrinth.raids.join.messages.log ( 'Trying to join ' + idrinth.raids.list[key].name + '\'s ' + idrinth.raids.list[key].raid );
-                    }
-                    try {
-                        idrinth.raids.joined[key] = idrinth.raids.list[key];
-                        delete idrinth.raids.list[key];
-                    } catch ( e1 ) {
-                    }
-                }
-            },
-            byFrame: {
-                windows: [ null, null, null ],
-                timeouts: [ ],
-                'do': function ( key, link ) {
-                    'use strict';
-                    for (var count = 0; count < idrinth.settings.windows; count++) {
-                        if ( idrinth.raids.join.byFrame.windows[count] === null || idrinth.raids.join.byFrame.windows[count] === undefined ) {
-                            idrinth.raids.join.byFrame.windows[count] =
-                                    idrinth.ui.buildElement ( { type: 'iframe', id: 'join-' + key, attributes: [
-                                            { name: 'src', value: link },
-                                            { name: 'sandbox', value: 'allow-forms allow-pointer-lock allow-popups allow-same-origin allow-scripts' },
-                                            { name: 'style', value: 'position:absolute;top:' + count + 'px;left:0;width:1px;height:1px;z-index:-100;' },
-                                            { name: 'onload', value: 'try{event.stopPropagation();}catch(e){}window.setTimeout(){function(){idrinth.raids.join.byFrame.remove(' + count + ',\'' + key + '\');},1234);' },
-                                            { name: 'onunload', value: 'try{event.stopPropagation();}catch(e){}' }
-                                        ] } );
-                            idrinth.raids.join.byFrame.timeouts[count] = eval ( 'window.setTimeout(function(){idrinth.raids.join.byFrame.remove(' + count + ',\'' + key + '\');},30000);' );
-                            idrinth.ui.body.appendChild ( idrinth.raids.join.byFrame.windows[count] );
-                            idrinth.raids.join.messages.trying ( key );
-                            return;
-                        }
-                    }
-                    idrinth.raids.list[key].joined = false;
-                },
-                remove: function ( count, key ) {
-                    try {
-                        window.clearTimeout ( idrinth.raids.join.byFrame.timeouts[count] );
-                    } catch ( e0 ) {
-                    }
-                    try {
-                        idrinth.raids.join.byFrame.windows[count] = null;
-                    } catch ( e1 ) {
-                    }
-                    try {
-                        idrinth.ui.removeElement ( 'join-' + key );
-                    } catch ( e2 ) {
-                    }
-                }
-            },
-            'do': function ( ) {
-                'use strict';
-                var added = 0;
-                for (var key in idrinth.raids.list) {
-                    if ( !idrinth.raids.list[key].joined ) {
-                        added++;
-                        idrinth.raids.join.servers.postLink ( key );
-                        idrinth.raids.list[key].joined = true;
-                        if ( typeof idrinth.raids.join.servers[idrinth.realSite] === 'function' ) {
-                            idrinth.raids.join.servers[idrinth.realSite] ( key );
-                        }
-                    }
-                    if ( added > 99 || ( idrinth.platform == 'facebook' && added >= idrinth.settings.windows ) ) {
-                        return true;
-                    }
-                }
-                return false;
-            },
-            process: function ( ) {
-                'use strict';
-                if ( !idrinth.raids.join.do ( ) && Date.now ( ) - 60000 > idrinth.raids.requested ) {
-                    idrinth.raids.requested = Date.now ( );
-                    idrinth.raids.import ( );
-                }
-            }
-        },
-        start: function ( ) {
-            'use strict';
-            idrinth.raids.restartInterval ();
         }
     },
     startInternal: function () {
@@ -2137,6 +1367,14 @@ var idrinth = {
             return;
         }
         idrinth.startInternal ();
+    },
+    /**
+     * @todo replace with a properly styled elements
+     * @param String text
+     * @returns Null
+     */
+    alert: function(text) {
+        window.alert(text);
     }
 };
 window.setTimeout ( function ( ) {
